requirements.txt
pinecone-client==2.2.1
langchain==0.0.174
unstructured==0.6.8
tabulate==0.9.0
pdf2image==1.16.3
pypdf==3.8.1
pdfminer.six==20221105
pdfminer==20191125
python_dotenv==1.0.0